- Refactored out page and util methods - wiki now exports a wiki enclosure that negates the need for classes, yay!
